fix(dashboard): make FeeFlowChart tooltip fully opaque for readability
Switch Recharts tooltip to an opaque card background and force wrapper opacity to 1.
Harden with CSS overrides and inline styles to prevent translucency during animations
and in dark mode. Also align global tooltip variables to opaque values.

Unrelated: widen whitelist epoch type to u32 in rust binding; add test addresses.

+++ crates/chainio/src/taiko/preconf_whitelist.rs
@@ -62,7 +62,7 @@ impl TaikoPreconfWhitelist {
         address: Address,
         epoch_timestamp: Timestamp,
     ) -> ContractResult<bool> {
-        match self.0.isOperatorActive(address, epoch_timestamp).call().await {
+        match self.0.isOperatorActive(address, epoch_timestamp as u32).call().await {
             Ok(result) => Ok(result),
             Err(err) => {
                 // Patch: handle the old whitelist contract implementation
@@ -119,7 +119,7 @@ sol! {
         /// @param operator The address to check.
         /// @param epochTimestamp The timestamp of the epoch to check.
         /// @return True if the address is an active operator in the given epoch.
-        function isOperatorActive(address operator, uint64 epochTimestamp) external view returns (bool);
+        function isOperatorActive(address operator, uint32 epochTimestamp) external view returns (bool);
 
         /// @notice The OLD whitelist method for checking if an address is whitelisted.
         /// @dev This is kept for backwards compatibility.
